requirements.txt
accelerate==1.12.0
albucore==0.0.24
albumentations==2.0.8
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.4
decord==0.6.0
easydict==1.13
einops==0.8.2
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.129.0
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
MarkupSafe==2.1.5
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
narwhals==2.15.0
networkx==3.4.2
numpy==2.2.6
ollama==0.6.1
onnx==1.20.1
onnxruntime-gpu==1.23.2
open_clip_torch==3.2.0
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.3
psutil==7.2.2
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
starlette==0.52.1
streamlit==1.52.2
stringzilla==4.6.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
tokenizers==0.15.2
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.1
transformers==4.39.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
wcwidth==0.2.14
win32_setctime==1.2.0
